requirements.txt
requests
user-agent
beautifulsoup4
python-dotenv
